Finished integrating new result parsing.
