+++ lib/pavilion/plugins/results/regex.py
@@ -4,6 +4,7 @@ import re
 import sre_constants
 
 import pavilion.result.base
+import pavilion.result.common
 import yaml_config as yc
 from pavilion.result import parsers
 
@@ -46,7 +47,7 @@ class Regex(parsers.ResultParser):
         try:
             kwargs['regex'] = re.compile(kwargs['regex'])
         except (ValueError, sre_constants.error) as err:
-            raise pavilion.result.base.ResultError(
+            raise pavilion.result.common.ResultError(
                 "Invalid regular expression: {}".format(err))
 
         return kwargs
@@ -56,9 +57,6 @@ class Regex(parsers.ResultParser):
         cregex = re.compile(regex)
 
         line = file.readline()
-
-        print('re matching', regex, line)
-
         match = cregex.search(line)
 
         if match is None:
